Move the SDK and platform-specific tools out of the repo
Place the Nordic programming tools and SDK outside the git repository and use environment variables to point to them.

This prevents having to maintain files in the local directory tree which aren't under git.

+++ firmware/update.sh
@@ -1,9 +1,11 @@
 #!/bin/bash
+: "${NRFJ_BIN:?Need to set NRFJ_BIN env variable (see project README)}"
+: "${SDK_ROOT:?Need to set SDK_ROOT env variable (see project README)}"
 
 cp manbearpig/_build/nrf52832_xxaa.hex .
-cp ../nRF5x/components/softdevice/s132/hex/s132_nrf52_3.0.0_softdevice.hex .
+cp $SDK_ROOT/components/softdevice/s132/hex/s132_nrf52_3.0.0_softdevice.hex .
 
-../bin/mergehex/mergehex -m nrf52832_xxaa.hex s132_nrf52_3.0.0_softdevice.hex -o andnxor_dc25_1.6.hex
+$NRFJ_BIN/mergehex/mergehex -m nrf52832_xxaa.hex s132_nrf52_3.0.0_softdevice.hex -o andnxor_dc25_1.6.hex
 
 rm nrf52832_xxaa.hex
 rm s132_nrf52_3.0.0_softdevice.hex
